Delete feed directory + a couple more py 2 -> 3 fixes
The RSS feed Python code is ancient and converting it was non-trivial so I'm
deleting it for now. There are a few subscribers (grep rss
/var/log/nginx/*access.log) but we need to be a little ruthless with where we
spend our time.

https://github.com/sphinx-contrib/yasfb looks like it might be a little more
recent and they've done some Python 3 conversion work already, so it might be a
good place to start if we decide to resurrect RSS feeds.